test: get sent messages from FakeNetwork
