+++ packages/client/test/test-utils/fake/FakeNetworkNode.ts
@@ -12,7 +12,7 @@ export class FakeNetworkNode implements NetworkNodeStub {
 
     public readonly id: string
     readonly subscriptions: Set<StreamPartID> = new Set()
-    private readonly messageListeners: MessageListener[] = []
+    readonly messageListeners: MessageListener[] = []
     private readonly network: FakeNetwork
 
     constructor(opts: NetworkNodeOptions, network: FakeNetwork) {
@@ -53,22 +53,7 @@ export class FakeNetworkNode implements NetworkNodeStub {
     }
 
     publish(msg: StreamMessage): void {
-        /*
-         * This serialization+serialization is needed in test/integration/Encryption.ts
-         * as it expects that the EncryptedGroupKey format changes in the process.
-         * TODO: should we change the serialization or the test? Or keep this hack?
-         */
-        const serialized = msg.serialize()
-        this.network.getNodes()
-            .forEach(async (networkNode) => {
-                if (networkNode.subscriptions.has(msg.getStreamPartID())) {
-                    networkNode.messageListeners.forEach((listener) => {
-                        // return a clone as client mutates message when it decrypts messages
-                        const deserialized = StreamMessage.deserialize(serialized)
-                        listener(deserialized)
-                    })
-                }
-            })
+        this.network.sendMessage(msg)
     }
 
     // eslint-disable-next-line class-methods-use-this
